+++ ext/cpp_polygon_finder/PolygonFinder/src/polygon/bitmaps/RawBitmap.h
@@ -32,6 +32,7 @@ class RawBitmap : public Bitmap {
   void fill(unsigned char r, unsigned char g, unsigned char b, unsigned char a = 255);
   void draw_line(int x0, int y0, int x1, int y1, unsigned char r, unsigned char g, unsigned char b, unsigned char a = 255);
   void draw_filled_rectangle(int x, int y, int w, int h, unsigned char r, unsigned char g, unsigned char b, unsigned char a = 255);
+  std::unique_ptr<RawBitmap> detach();
 
  protected:
   std::unique_ptr<unsigned char[]> image;