+++ lib/contrek/finder/result.rb
@@ -8,6 +8,12 @@ module Contrek
       def points
         polygons
       end
+
+      def sort_polygons!
+        polygons.sort_by! do |polygon|
+          [polygon[:bounds][:min_y], polygon[:bounds][:min_x]]
+        end
+      end
     end
   end
 end